conversational-ai-platform/pyproject.toml
[project]
name = "conversational-ai-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.3.0",
    "google-auth",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-generativeai>=0.8.4",
    "litellm>=1.63.3",
]
